Support -os-provides-dlopen
Add a new C function - arch_read_linkage_entry - that backs out the true
foreign address of the corresponding symbol from the data in the linkage table.

Add a Lisp function - FIND-LINKAGE-TABLE-FOREIGN-SYMBOL-ADDRESS - that uses
this new C function to return the address, while accounting for undefined
symbols (so as to not accidentally return a reference to a guard page).

Teach FIND-FOREIGN-SYMBOL-ADDRESS to use
FIND-LINKAGE-TABLE-FOREIGN-SYMBOL-ADDRESS when libdl is not available. Replace
all calls to FIND-DYNAMIC-FOREIGN-SYMBOL-ADDRESS with
FIND-FOREIGN-SYMBOL-ADDRESS as the former is not available on
-os-provides-dlopen and the latter correctly falls back the the best option
based on *FEATURES*.